refactor(bridge-exec): split bridge proof executor into own module
